Fix circular imports in forms package. This resolves a problem where Rollup's watch mode wasn't working.
