+++ packages/forms/src/builder/builder.test.ts
@@ -2,9 +2,9 @@ import { describe, expect, it } from 'vitest';
 
 import { type Pattern, createForm, getPattern } from '../index.js';
 import { defaultFormConfig } from '../patterns/index.js';
-import { type FieldsetPattern } from '../patterns/fieldset/index.js';
+import { type FieldsetPattern } from '../patterns/fieldset/config.js';
 import { type FormSummaryPattern } from '../patterns/form-summary.js';
-import { type InputPattern } from '../patterns/input/index.js';
+import { type InputPattern } from '../patterns/input/config.js';
 import { type PagePattern } from '../patterns/page/config.js';
 import { type PageSetPattern } from '../patterns/page-set/config.js';
 import { type RadioGroupPattern } from '../patterns/radio-group.js';
