+++ packages/forms/src/patterns/input/index.ts
@@ -1,12 +1,10 @@
 import { enLocale as message } from '@atj/common';
 
-import { Pattern, type PatternConfig } from '../../pattern.js';
+import { type PatternConfig } from '../../pattern.js';
 
-import { parseConfigData, type InputConfigSchema } from './config.js';
+import { type InputPattern, parseConfigData } from './config.js';
 import { createPrompt } from './prompt.js';
 import { type InputPatternOutput, parseUserInput } from './response.js';
-
-export type InputPattern = Pattern<InputConfigSchema>;
 
 export const inputConfig: PatternConfig<InputPattern, InputPatternOutput> = {
   displayName: message.patterns.input.displayName,
