+++ packages/forms/src/patterns/input/response.ts
@@ -3,7 +3,7 @@ import { z } from 'zod';
 import { ParseUserInput } from '../../pattern.js';
 import { safeZodParseToFormError } from '../../util/zod.js';
 
-import { type InputPattern } from './index.js';
+import { type InputPattern } from './config.js';
 
 const createSchema = (data: InputPattern['data']) => {
   const schema = z.string().max(data.maxLength);
