+++ packages/forms/src/types.ts
@@ -1,0 +1,21 @@
+import { type DocumentFieldMap } from './documents/types';
+import { type PatternId, type PatternMap } from './pattern';
+
+export type Blueprint = {
+  summary: FormSummary;
+  root: PatternId;
+  patterns: PatternMap;
+  outputs: FormOutput[];
+};
+
+export type FormSummary = {
+  title: string;
+  description: string;
+};
+
+export type FormOutput = {
+  data: Uint8Array;
+  path: string;
+  fields: DocumentFieldMap;
+  formFields: Record<string, string>;
+};
